+++ changelog.md
@@ -1,1 +1,1 @@
-- Clean small river.
+- Add some fallback for world init.